Add files via upload
Functions for finding games that a user is in and authenticating a user login. I wasn't sure what to return for the games, so I just returned the game name. Also, for login, since there's nothing to login to yet, I just printed out either "valid combination" or "invalid combination".